+++ tests/core/parameters_date_indexing/test_date_indexing.py
@@ -5,8 +5,6 @@ import numpy
 from openfisca_core.parameters import ParameterNode
 from openfisca_core.tools import assert_near
 
-from openfisca_core.model_api import *  # noqa
-
 LOCAL_DIR = os.path.dirname(os.path.abspath(__file__))
 
 parameters = ParameterNode(directory_path=LOCAL_DIR)